chore(kernel): move tests up one level
