+++ packages/kernel/test/reporter.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { Reporter } from '../../src/index';
+import { Reporter } from '../src/index';
 
 describe('The default Reporter', () => {
   it('creates an Error with a minimal message upon request', () => {
