--- packages/kernel/test/unit/util.ts
@@ -1,3 +1,0 @@
-import { _ } from '../../../../scripts/test-lib';
-
-export { _ };
