+++ packages/kernel/test/util.ts
@@ -1,0 +1,3 @@
+import { _ } from '../../../scripts/test-lib';
+
+export { _ };
